requirements.txt
discord.py
aiohttp
pytz
beautifulsoup4
jishaku
git+https://github.com/Rapptz/discord-ext-menus